Convert hook to regular function
