+++ apps/melding-form/src/app/(general)/_utils/getDocumentTitleOnError.test.ts
@@ -1,0 +1,24 @@
+import { getDocumentTitleOnError } from './getDocumentTitleOnError'
+
+describe('getDocumentTitleOnError', () => {
+  const originalDocTitle = 'Test title'
+  const t = (_: string, { count }: { count: number }) => `(${count} errors)`
+
+  it('returns original title when there are no validationErrors', () => {
+    const title = getDocumentTitleOnError(originalDocTitle, t, undefined)
+
+    expect(title).toBe(originalDocTitle)
+  })
+
+  it('returns error count label and original title when validationErrors are present', () => {
+    const validationErrors = [
+      { key: 'a', message: 'Error A' },
+      { key: 'b', message: 'Error B' },
+      { key: 'c', message: 'Error C' },
+    ] // 3 errors
+
+    const title = getDocumentTitleOnError(originalDocTitle, t, validationErrors)
+
+    expect(title).toBe('(3 errors) Test title')
+  })
+})
